Pull derived entities into button component
